Face-Recognite-AI-Edge-VietNam-main/requirements.txt
faiss-cpu
tensorflow
keras
onnxruntime
openvino
opencv-python
mtcnn
ultralytics
numpy
Pillow
av
pymongo
Flask
